StartupHack/requirements.txt
psycopg2-binary
uvicorn==0.29.0
fastapi==0.111.0
asyncio
langchain-core
langchain-openai
pydantic
python-dotenv
requests
ngrok==1.4.0
langchain>=0.1.0
langchain_community
pyngrok
asyncpg
pyngrok